Logging works. :sparkles:  Fixes Issues #2 and #17 :bug:

+++ src/Includes/Utilities/FS.hpp
@@ -8,6 +8,7 @@
 #include <cstdlib>
 #include <ctime>
 #include "Utilities.hpp"
+#include "Logging.hpp"
 
 namespace IO
 {
@@ -53,16 +54,22 @@ size_t CalculateNextSeed()
 
 void PrepareNextIter(const std::string &paramsName, const size_t &iter)
 {
-    // std::cout << "Start of PrepareNexIter " << std::endl;
     using boost::filesystem::copy_file;
+    using boost::filesystem::exists;
     const std::string ext = ".dat";
 
     copy_file("hybNextUp.dat", std::string("hybUp") + std::to_string(iter + 1) + ext);
     copy_file("selfUp.dat", std::string("selfUp") + std::to_string(iter) + ext);
     copy_file("greenUp.dat", std::string("greenUp") + std::to_string(iter) + ext);
-    copy_file("Obs.json", std::string("Obs") + std::to_string(iter) + ".json");
-    copy_file("upd.meas.json", std::string("upd.meas") + std::to_string(iter) + ".json");
-    copy_file("gtau.dat", std::string("g0tau") + std::to_string(iter) + ext);
+    if (exists("Obs.json") && Logging::LevelIsTrace())
+    {
+        copy_file("Obs.json", std::string("Obs") + std::to_string(iter) + ".json");
+    }
+    if (exists("gtau.dat") && Logging::LevelIsTrace())
+    {
+
+        copy_file("gtau.dat", std::string("g0tau") + std::to_string(iter) + ext);
+    }
 
 #ifdef AFM
     copy_file("hybNextDown.dat", std::string("hybDown") + std::to_string(iter + 1) + ext);
@@ -70,7 +77,7 @@ void PrepareNextIter(const std::string &paramsName, const size_t &iter)
     copy_file("greenDown.dat", std::string("greenDown") + std::to_string(iter) + ext);
 #endif
 
-    std::string fname = paramsName + std::to_string(iter) + std::string(".json");
+    const std::string fname = paramsName + std::to_string(iter) + std::string(".json");
     std::ifstream fin(fname);
     Json params;
     fin >> params;
